DigiCaddie/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
boto3==1.34.64
botocore==1.34.64
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==4.2.7
django-bootstrap-icons==0.8.7
django-bootstrap-v5==1.0.11
django-storages==1.14.2
idna==3.6
jmespath==1.0.1
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
